+++ CHANGELOG.md
@@ -3,10 +3,11 @@
 ## Unreleased
 * Docker
 
-## 0.2.2 (Unreleased)
+## 0.2.2 (2020-10-01)
 * Excluded actions no longer show up in results (Fixes #106)
 * Fixed issue where `*:*` policy would break results due to how the Service Wildcard finding was implemented (Fixes #109)
 * Credentials Exposure and Service Wildcard now show up in the data file results
+* Exclusions are now applied earlier in the scan
 
 ## 0.2.1 (9/25/2020)
 * Fixes issue where Inline Policies were showing up as findings even when they were attached to excluded IAM principals. Fixes #104